backend/requirements.txt
Flask
Flask-CORS
requests
pydantic
python-dotenv
